Extract detectCombinedPackage to shared utility
Deduplicates the function that was copied in both cli.ts and generate.ts
into a single shared module at commands/utils/detect-package.ts.
