+++ packages/cli/commands/utils/generate.ts
@@ -1,23 +1,6 @@
 import fs from 'node:fs';
 import path from 'node:path';
-
-/**
- * Detects whether the combined 'monorise' package is installed by walking up
- * the directory tree. This handles monorepo setups where dependencies are
- * hoisted to the root node_modules.
- */
-function detectCombinedPackage(startDir: string): boolean {
-  let dir = startDir;
-  while (true) {
-    if (fs.existsSync(path.join(dir, 'node_modules', 'monorise'))) {
-      return true;
-    }
-    const parent = path.dirname(dir);
-    if (parent === dir) break; // reached filesystem root
-    dir = parent;
-  }
-  return false;
-}
+import { detectCombinedPackage } from './detect-package';
 
 function kebabToCamel(str: string): string {
   return str.replace(/-([a-z])/g, (_, letter) => letter.toUpperCase());
